arduino-modbus-rtu-tcp-gateway.ino: changed deprecated header to non-deprecated one. 03-modbus-rtu.ino: Bugfix: changed rxNdx from uint8_t to uint16_t, because the following "if (rxNdx < MODBUS_SIZE)" would have been always true. 05-pages.ino: initialised byte value with 0 because of "uninitialized variable"-warning from compiler.

+++ arduino-modbus-rtu-tcp-gateway/03-modbus-rtu.ino
@@ -123,7 +123,7 @@ void sendSerial() {
 }
 
 void recvSerial() {
-  static uint8_t rxNdx = 0;
+  static uint16_t rxNdx = 0;
   static byte serialIn[MODBUS_SIZE];
   while (mySerial.available() > 0) {
     byte b = mySerial.read();
